Hold every nutrient figure as a Decimal (#22)
* Pin that a shard round trip reproduces it byte for byte

* Serialize a Decimal, and emit through the record serializer

The shards and the wire carry the same figures, so they get the same
serializer. A Decimal already carries the digits and exponent this format is
written in terms of, so `_digits` no longer recovers them from a `repr`, and
`json.dumps` -- which cannot write a Decimal at all -- leaves the JSON path.

* Hold every nutrient figure as a Decimal

A figure on a label is a decimal that was printed, not a measured binary
quantity. Read as a float it scales into noise -- 0.7 g of fat over 40 g came
out 0.27999999999999997 -- which a record and then a recipe file carried
verbatim.

`parse_float=Decimal` keeps the digits a line states, so a read-modify-write
round trip is exact by construction rather than by float formatting being
reproduced. Restating multiplies before dividing, and dividing by a power of
ten becomes a shift, so the only rounding left is the one an inexact division
genuinely needs. The single float boundary is named: `_MAX_FIGURE`, the
largest figure a JSON consumer reading doubles can hold.

* Refuse a float where a figure belongs

The serializer had a float branch only because a float has no digits of its
own and `repr` had to supply them. Nothing reaches it with one now, so
removing it turns a stray float into a TypeError rather than a rewritten
shard.

* Bound both sides of a restatement, not only its result

A weight past the largest double is finite to a Decimal, so it reached the
arithmetic and overflowed the context instead of being refused.

+++ src/pantry/commands/lookup.py
@@ -1,11 +1,14 @@
 """`pantry lookup` — exact composite-identity lookup."""
 
+from decimal import Decimal
+
 import click
-from agentcli import emit, json_option
+from agentcli import json_option
 
 from pantry.commands import grams_option
 from pantry.commands.describe import describe
 from pantry.local import as_result
+from pantry.output import emit
 from pantry.products import PRODUCT_SOURCES, rescale
 from pantry.session import deps, guard, wants_json
 
@@ -26,7 +29,7 @@ def lookup(
     ctx: click.Context,
     source: str,
     product_id: str,
-    grams: float | None,
+    grams: Decimal | None,
     json_output: bool,
 ) -> None:
     """Find exactly SOURCE and PRODUCT_ID, with no fuzz and no network.
